+++ CHANGES.md
@@ -1,5 +1,7 @@
-# unreleased
 
+# Unreleased
+
+* Enable dynamic definition of Asset reader in MultiBaseReader
 * better error message for `TileOutsideBounds` errors (author @abarciauskas-bgse, https://github.com/cogeotiff/rio-tiler/pull/712)
 
 # 6.6.1 (2024-05-17)